Move GetClasses into its own file
